Refactor convertReadingValueType method
