+++ internal/driver/driver.go
@@ -522,89 +522,3 @@ func checkValueInRange(profileValueType string, readingValueType string, reading
 
 	return isValid
 }
-
-func convertReadingValueType(profileValueType string, readingValueType string, reading interface{}) interface{} {
-	if readingValueType == "int" {
-		switch profileValueType {
-		case "int8":
-			reading = int8(reading.(int64))
-		case "int16":
-			reading = int16(reading.(int64))
-		case "int32":
-			reading = int32(reading.(int64))
-		case "int64":
-			reading = int64(reading.(int64))
-		case "uint8":
-			reading = uint8(reading.(int64))
-		case "uint16":
-			reading = uint16(reading.(int64))
-		case "uint32":
-			reading = uint32(reading.(int64))
-		case "uint64":
-			reading = uint64(reading.(int64))
-		}
-	} else if readingValueType == "uint" {
-		switch profileValueType {
-		case "int8":
-			reading = int8(reading.(uint64))
-		case "int16":
-			reading = int16(reading.(uint64))
-		case "int32":
-			reading = int32(reading.(uint64))
-		case "int64":
-			reading = int64(reading.(uint64))
-		case "uint8":
-			reading = uint8(reading.(uint64))
-		case "uint16":
-			reading = uint16(reading.(uint64))
-		case "uint32":
-			reading = uint32(reading.(uint64))
-		case "uint64":
-			reading = uint64(reading.(uint64))
-		}
-	} else if readingValueType == "float64" {
-		switch profileValueType {
-		case "int8":
-			reading = int8(reading.(float64))
-		case "int16":
-			reading = int16(reading.(float64))
-		case "int32":
-			reading = int32(reading.(float64))
-		case "int64":
-			reading = int64(reading.(float64))
-		case "uint8":
-			reading = uint8(reading.(float64))
-		case "uint16":
-			reading = uint16(reading.(float64))
-		case "uint32":
-			reading = uint32(reading.(float64))
-		case "uint64":
-			reading = uint64(reading.(float64))
-		case "float32":
-			reading = float32(reading.(float64))
-		}
-	} else if readingValueType == "float32" {
-		switch profileValueType {
-		case "int8":
-			reading = int8(reading.(float32))
-		case "int16":
-			reading = int16(reading.(float32))
-		case "int32":
-			reading = int32(reading.(float32))
-		case "int64":
-			reading = int64(reading.(float32))
-		case "uint8":
-			reading = uint8(reading.(float32))
-		case "uint16":
-			reading = uint16(reading.(float32))
-		case "uint32":
-			reading = uint32(reading.(float32))
-		case "uint64":
-			reading = uint64(reading.(float32))
-		case "float64":
-			reading = float64(reading.(float32))
-		}
-	}
-
-	return reading
-}
